requirements.txt
opencv-python
imutils
imagededup
scipy
pyiqa
numpy<2
Pillow
